Move dynamic fee retriaval from btc/transactions to service utility
