+++ src/utilities/api/btc/service.js
@@ -1,7 +1,6 @@
-// import config from '../../../../btc.config';
+import config from '../../../../btc.config';
 import liskConfig from '../../../../lsk.config';
 
-// eslint-disable-next-line import/prefer-default-export
 export const getPriceTicker = () => new Promise(async (resolve, reject) => {
   try {
     const response = await fetch(`${liskConfig.serviceURL}/api/getPriceTicker`, liskConfig.requestOptions);
@@ -13,6 +12,25 @@ export const getPriceTicker = () => new Promise(async (resolve, reject) => {
       resolve({
         EUR: String(BTC.EUR),
         USD: String(BTC.USD),
+      });
+    } else {
+      reject(json);
+    }
+  } catch (error) {
+    reject(error);
+  }
+});
+
+export const getDynamicFees = () => new Promise(async (resolve, reject) => {
+  try {
+    const response = await fetch(config.minerFeesURL);
+    const json = await response.json();
+
+    if (response.ok) {
+      resolve({
+        low: json.hourFee,
+        medium: json.halfHourFee,
+        high: json.fastestFee,
       });
     } else {
       reject(json);
