+++ src/utilities/api/btc/transactions.test.js
@@ -1,6 +1,7 @@
 import bitcoin from 'bitcoinjs-lib';
 import fetchMock from 'fetch-mock';
 import * as transactions from './transactions';
+import * as service from './service';
 import config from '../../../../btc.config';
 
 const passphrase = 'say width dwarf confirm rule party pact iron edge dignity wish direct';
@@ -154,19 +155,6 @@ const response = {
       },
     ],
   },
-  getMinerFees: {
-    hourFee: 20,
-    halfHourFee: 40,
-    fastestFee: 80,
-  },
-};
-
-const normalizedData = {
-  minerFees: {
-    low: 20,
-    medium: 40,
-    high: 80,
-  },
 };
 
 describe('api/btc/transactions', () => {
@@ -282,37 +270,6 @@ describe('api/btc/transactions', () => {
     });
   });
 
-  describe('getMinerFees', () => {
-    beforeEach(() => fetchMock.reset());
-
-    it('resolves correctly', async () => {
-      fetchMock.once('*', response.getMinerFees);
-      const result = await transactions.getMinerFees();
-      expect(result).toEqual(normalizedData.minerFees);
-    });
-
-    it('handles non-500 errors', async () => {
-      const errorResponse = { message: 'Error' };
-      fetchMock.once('*', { status: 400, body: errorResponse });
-
-      try {
-        await transactions.getMinerFees();
-      } catch (error) {
-        expect(error).toEqual(errorResponse);
-      }
-    });
-
-    it('handles errors', async () => {
-      fetchMock.once('*', { throws: new TypeError('Failed to fetch') });
-
-      try {
-        await transactions.getMinerFees();
-      } catch (error) {
-        expect(error).toBeTruthy();
-      }
-    });
-  });
-
   describe('getUnspentOuts', () => {
     beforeEach(() => fetchMock.reset());
 
@@ -348,8 +305,12 @@ describe('api/btc/transactions', () => {
     beforeAll(() => {
       config.network = bitcoin.networks.testnet;
 
-      transactions.getMinerFees = jest.fn();
-      transactions.getMinerFees.mockResolvedValue(normalizedData.minerFees);
+      service.getDynamicFees = jest.fn();
+      service.getDynamicFees.mockResolvedValue({
+        low: 20,
+        medium: 40,
+        high: 80,
+      });
 
       transactions.getUnspentOuts = jest.fn();
       transactions.getUnspentOuts.mockResolvedValue(response.getUnspentOuts.unspent_outputs);
